move meta to policy service
