+++ pkg/microservice/aslan/core/collaboration/service/collaboration_instance.go
@@ -39,6 +39,7 @@ import (
 	workflowservice "github.com/koderover/zadig/pkg/microservice/aslan/core/workflow/service/workflow"
 	"github.com/koderover/zadig/pkg/setting"
 	"github.com/koderover/zadig/pkg/shared/client/policy"
+	"github.com/koderover/zadig/pkg/types"
 )
 
 type GetCollaborationUpdateResp struct {
@@ -377,19 +378,19 @@ func buildPolicybindingName(uid, policyName, projectName string) string {
 
 func syncPolicy(updateResp *GetCollaborationUpdateResp, projectName, identityType, userName, uid string,
 	logger *zap.SugaredLogger) error {
-	var policies []*policy.Policy
+	var policies []*types.Policy
 	var policyBindings []*policy.PolicyBinding
 	for _, mode := range updateResp.New {
-		var rules []*policy.Rule
+		var rules []*types.Rule
 
 		policyName := buildPolicyName(projectName, mode.Name, identityType, userName)
 
 		for _, workflow := range mode.Workflows {
-			rules = append(rules, &policy.Rule{
+			rules = append(rules, &types.Rule{
 				Verbs:     workflow.Verbs,
 				Kind:      "resource",
 				Resources: []string{string(config2.ResourceTypeWorkflow)},
-				MatchAttributes: []policy.MatchAttribute{
+				MatchAttributes: []types.MatchAttribute{
 					{
 						Key:   "policy",
 						Value: buildLabelValue(projectName, mode.Name, identityType, userName, string(config2.ResourceTypeWorkflow), workflow.Name),
@@ -398,11 +399,11 @@ func syncPolicy(updateResp *GetCollaborationUpdateResp, projectName, identityTyp
 			})
 		}
 		for _, product := range mode.Products {
-			rules = append(rules, &policy.Rule{
+			rules = append(rules, &types.Rule{
 				Verbs:     product.Verbs,
 				Kind:      "resource",
 				Resources: []string{string(config2.ResourceTypeProduct)},
-				MatchAttributes: []policy.MatchAttribute{
+				MatchAttributes: []types.MatchAttribute{
 					{
 						Key:   "policy",
 						Value: buildLabelValue(projectName, mode.Name, identityType, userName, string(config2.ResourceTypeProduct), product.Name),
@@ -410,7 +411,7 @@ func syncPolicy(updateResp *GetCollaborationUpdateResp, projectName, identityTyp
 				},
 			})
 		}
-		policies = append(policies, &policy.Policy{
+		policies = append(policies, &types.Policy{
 			Name:        policyName,
 			UpdateTime:  time.Now().Unix(),
 			Description: buildPolicyDescription(mode.Name, userName),
@@ -440,15 +441,15 @@ func syncPolicy(updateResp *GetCollaborationUpdateResp, projectName, identityTyp
 			return err
 		}
 	}
-	var updatePolicies []*policy.Policy
+	var updatePolicies []*types.Policy
 	for _, instance := range updateResp.UpdateInstance {
-		var rules []*policy.Rule
+		var rules []*types.Rule
 		for _, workflow := range instance.Workflows {
-			rules = append(rules, &policy.Rule{
+			rules = append(rules, &types.Rule{
 				Verbs:     workflow.Verbs,
 				Kind:      "resource",
 				Resources: []string{string(config2.ResourceTypeWorkflow)},
-				MatchAttributes: []policy.MatchAttribute{
+				MatchAttributes: []types.MatchAttribute{
 					{
 						Key:   "policy",
 						Value: buildLabelValue(projectName, instance.CollaborationName, identityType, userName, string(config2.ResourceTypeWorkflow), workflow.BaseName),
@@ -457,11 +458,11 @@ func syncPolicy(updateResp *GetCollaborationUpdateResp, projectName, identityTyp
 			})
 		}
 		for _, product := range instance.Products {
-			rules = append(rules, &policy.Rule{
+			rules = append(rules, &types.Rule{
 				Verbs:     product.Verbs,
 				Kind:      "resource",
 				Resources: []string{string(config2.ResourceTypeProduct)},
-				MatchAttributes: []policy.MatchAttribute{
+				MatchAttributes: []types.MatchAttribute{
 					{
 						Key:   "policy",
 						Value: buildLabelValue(projectName, instance.CollaborationName, identityType, userName, string(config2.ResourceTypeProduct), product.BaseName),
@@ -469,7 +470,7 @@ func syncPolicy(updateResp *GetCollaborationUpdateResp, projectName, identityTyp
 				},
 			})
 		}
-		updatePolicies = append(updatePolicies, &policy.Policy{
+		updatePolicies = append(updatePolicies, &types.Policy{
 			Name:        instance.PolicyName,
 			Description: buildPolicyDescription(instance.CollaborationName, userName),
 			UpdateTime:  time.Now().Unix(),
