+++ pkg/microservice/policy/core/handler/policy_meta_registration.go
@@ -23,20 +23,6 @@ import (
 	internalhandler "github.com/koderover/zadig/pkg/shared/handler"
 )
 
-func CreateOrUpdatePolicyRegistration(c *gin.Context) {
-	ctx := internalhandler.NewContext(c)
-	defer func() { internalhandler.JSONResponse(c, ctx) }()
-
-	args := &service.PolicyMeta{}
-	if err := c.ShouldBindJSON(args); err != nil {
-		ctx.Err = err
-		return
-	}
-	args.Resource = c.Param("resourceName")
-
-	ctx.Err = service.CreateOrUpdatePolicyRegistration(args, ctx.Logger)
-}
-
 func GetPolicyRegistrationDefinitions(c *gin.Context) {
 	ctx := internalhandler.NewContext(c)
 	defer func() { internalhandler.JSONResponse(c, ctx) }()
