+++ pkg/microservice/policy/core/handler/router.go
@@ -103,11 +103,6 @@ func (*Router) Inject(router *gin.RouterGroup) {
 		bundles.GET("/:name", DownloadBundle)
 	}
 
-	policyRegistrations := router.Group("policymetas")
-	{
-		policyRegistrations.PUT("/:resourceName", CreateOrUpdatePolicyRegistration)
-	}
-
 	policyDefinitions := router.Group("policy-definitions")
 	{
 		policyDefinitions.GET("", GetPolicyRegistrationDefinitions)
